Refactor SgOmpParallelStatement to SgUpirSpmdStatement
In src/roseSupport/stringify.C and stringify.h, and some files outside the src folder, OmpParallelStatement is still used, but it doesn't affect the AST constuction and transformation.
At the moment, we leave them as-is.

scripts/restringify can regenerate stringify.C and stringify.h, but because we remove many files from original ROSE, the new regenerated files do not work properly and will cause a building failure. Therefore, we do not touch these two files for now.